% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie.r
\name{colorkey}
\alias{colorkey}
\title{Function to create a color key}
\usage{
colorkey(x0, y0, x1, y1, label.from, label.to, tick.range = c(0, 1),
  tick.marks = c(0, 0.25, 0.5, 0.75, 1), cols = c("white", "steelblue"),
  nlevels = 100, main = NULL)
}
\arguments{
\item{x0, x1, y0, y1}{Numeric scalars. Coordinates of the lower left and
upper right points where the color key will be drawn as proportion of
the plotting region.}

\item{label.from, label.to}{Character scalar. Labels of the lower and
upper values of the color key.}

\item{tick.range, tick.marks}{Numeric vectors specifying the range and
the tickmarks respectively.}

\item{cols}{Character scalar. Colors specifications to create the
color palette.}

\item{nlevels}{Integer scalar. Number of levels to extrapolate.}

\item{main}{Character scalar. Title of the colorkey.}
}
\description{
Function to create a color key
}
